+++ R/Means_DDCt.r
@@ -3,7 +3,7 @@
 #' @description
 #' Performs relative expression (fold change) analysis based on the
 #' Delta Delta Ct (ddCt) methods using a fitted model object produced by
-#' \code{ANOVA_DCt()}, \code{ANOVA_DDCt()} or \code{REPEATED_DDCt()}.
+#' \code{ANOVA_DCt()} or \code{ANOVA_DDCt()}.
 #'
 #' @details
 #' The \code{Means_DDCt} function performs pairwise comparisons of relative expression values fo all combinations using
@@ -26,7 +26,7 @@
 #'
 #' @param model
 #' A fitted model object (typically an \code{lmer} or \code{lm} object)
-#' created by \code{ANOVA_DCt()}, \code{ANOVA_DDCt()} or \code{REPEATED_DDCt()}.
+#' created by \code{ANOVA_DCt()}, \code{ANOVA_DDCt()}.
 #'
 #' @param specs
 #' A character string or character vector specifying the predictors or